+++ packages/usb_iso_cli/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.3.0
 
 - `make --iso Win11.iso --iso ubuntu.iso` writes a GRUB multiboot USB
+- `add --iso` copies another image onto an existing multiboot USB
+- `refresh` rebuilds the GRUB menu from files already on the stick
 
 ## 1.2.0